+++ srcs/models/gemma2/gemma_update_mask.cuh
@@ -4,7 +4,7 @@
 #include "tensor.hpp"
 
 void cu_tensor_gemma_update_mask_f32(
-	const Tensor<int32, CUDA>& attention_mask,
+	const Tensor<float32, CUDA>& attention_mask,
 	const Tensor<int32, CUDA>& cache_position,
 	const int32 trg_len,
 	Tensor<float32, CUDA>& yt);
